requirements_common.txt
Flask==3.0.3
gunicorn==23.0.0
psycopg2==2.9.9
requests==2.32.3
authlib==1.3.1
pyjwt==2.9.0
